Fixed aggregateCumulative
aggregateCumulative was originally using Pipes.Prelude.head over
Pipes.next to get the first point. Due to the monadic nature of the
data, this would cause the underlying query to be repeated and lead
to unpredicatable data. Since Pipes.next works in the same pull-based
manner as >-> and most other Pipes constructs, it is well-behaved in
line with our expectations.

Tests did not pick this up as the test data was pure.